requirements.txt
django==3.2.1
djangorestframework==3.12.4
python-dotenv==0.15.0
psycopg2-binary==2.8.6
requests==2.25.1
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
